// src/types.ts
export interface LocationMetadata {
  placeName: string;
  formattedAddress: string;
  city?: string;
  district?: string;
  state?: string;
  country: string;
  lat: number;
  lon: number;
  boundingBox?: [string, string, string, string];
  osmId?: string | number;
  type?: string;
}